Changed movies and cast members

+++ MovieCollectionManagement/MovieCollectionManagement/src/main/java/pl/polsl/moviecollectionmanagement/dtos/MovieDto.java
@@ -24,7 +24,7 @@ public class MovieDto {
 
     private String description;
 
-    private Integer boxOffice;
+    private Float boxOffice;
 
     private Integer seasonNumber;
